enbale visiable map and login_form
