+++ Sources/Server/Client/Debug/Client_c.tlog/Client_c.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
+Debug|Win32|C:\Users\dienv\Desktop\LTM_TUNGBT_20172\Sources\Server\|
